Added fields for individual lines of text related to every rating. This lets us put a sentence or two next to the stars giving our reasons. Added migrations and fields to Admin.
Fleshed out Review view -- this is pretty complete at this point (in terms of getting things on the page) and doesn't look bad at all. We may want to collapse the tasks at some point and provide an expand button. Comments will probably also have to be added, as well as the user-facing rating component.

Added a block in the base file at the top of the page for full-width content. May use this on the index page.