Fleshed out "More" template.
Added a new CustomTag model so that we can run the tagging stuff through our own model. For now all this does is lowercase the tags and provides a get_absolute_url method (TODO). Added that model to the admin as well. This should make the tags more flexible long-term.

Migration for new models.