Modfied the search view to also be used as the 'more' view for a given model. That view is now pretty useless and may be removed.
Modified the search view to take a date range (hardcoded ranges).

Added "facets" to search results page. These are the date ranges and the model types. Added a renderer for that filter.

Modified some CSS so it looks okay. The design gets wonky if there's only one result and it doesn't have enough content to be the width of the center well (this will probably never happen in reality so I ignored it).