Added a manager to the models that returns only published items. Made the default "objects" in the CommonInfo model use the PublishedManager for returning results. This means that by default only published items will ever come back in a query set.
This introduced an issue where the admin interface wouldn't allow you unpublish an item without it disappearing forever. The CommonAdmin model was adjusted to override the default queryset and return all of the models using a new "all_objects" manager on the CommonInfo model.

This seems to work fine for now with at least one exception. In the templates if related foreign keys aren't published they'll still appear because right now I'm checking for existence, not published (and they don't go through a manager on foreign keys). So, if a ProductTask has a Task that's published and a DoucmentSet that isn't, both will appear but clicking on the DocumentSet will 404. This may have to be fixed, but it's hard to imagine a situation where it would arise intentionally.