Okay, this change touched lots of places and needs to be kept an eye on, but it was pretty straightforward:
It became apparent that overriding the default "objects" property on a model is BAD(tm). The fixes to get around that became greater than the advantage of using it to show only published items. So, the CommonInfo model was changed to have a standard objects property as well as a published_objects property. The views were all updated to use the published_objects property by default. The admin was changed around to stop overriding to use the old all_objects property.

In general the "published_objects" property should be used on anything public facing, and the standard 'objects' property should be used within the admin.