Made it so only members of the "Editors" group can publish.
This may be implemented in an insane way. Nothing I could do would get the formset to change in a sane way. I ended up copying the get_fieldsets method into every ModelAdmin because it wouldn't work properly in the CommonAdmin.

Also had to override how the list display worked with the editable.