Added filter for getting classname into template.
Moved 'ratings box' down the page slightly on reviews for approval. Others will have to be moved if this is where we want it.

Added a 'category finder' box to top of index page. This is a category dropdown and a search box for the review system.

Added the same box to the top of the search results page. Also included current search text in this box (ie 'showing results for: Outwit'). Does not include current facets.

Added a facet for "Category." This will always force the 'type' to be 'products.' It will retain the currently filtered date range and query.

Removed some extraneous mentions to old design 960 framework (ie: alpha, omega, grid_16, etc).

CSS to match above creations.