Moved Duke Logo down 5px for more spacing.
Added Category descriptions to Find A Tool Box when one is selected.

Removed other "tools" and replaced with "Coming Soon" -- Took away the hardcoded size of those black boxes, but we're going to want to add it back when we have more tools to prevent Jengaing

Changed Row Design Studios to white

Changed "active" color for facets to a light gray as opposed to green. The yellow was awfully garish, and the light blue barely showed up. The gray is the same as the dropshadows.