Added close.png file for close buttons (made it myself, so maybe we'll need to adjust it in the future.)
Added jquery tools script to base template (http://cdn.jquerytools.org/1.2.6/all/jquery.tools.min.js) -- this provides a light-weight overlay framework as well as some other jquery helpers if we need them in the future. Free and open source (creator doesn't believe in copyrights).

Added a "?" next to the category dropdown on both the home page and the search results page. This will pop open an overlay listing all of the categories and their descriptions.