Added a test to see if the League Gothic font loaded and adjust design accordingly.
Inside the Sanford building there's something wrong with the permissions on some of the PCs. The fonts will not load in various versions of Firefox and Internet Explorer. The fonts are embedded in a very standard way, and trying other embed styles hasn't worked (in fact, the fonts won't load even on example sites or major publications). What this does is check two divs to see if their size is the same, oen with League Gothic and one with Arial. Since League Gothic is larger, they'll fail this test unless the font didn't load (Arial is the fallback). If that fails, we add a class to the body and fix some styles so it's not horrible.